# contracts/openapi.json
{
  "components": {
    "schemas": {
      "BrowseFoldersIn": {
        "properties": {
          "auth_mode": {
            "enum": [
              "service_account",
              "oauth"
            ],
            "title": "Auth Mode",
            "type": "string"
          },
          "credential_ref": {
            "minLength": 1,
            "title": "Credential Ref",
            "type": "string"
          },
          "parent_id": {
            "default": "root",
            "title": "Parent Id",
            "type": "string"
          },
          "principal": {
            "minLength": 1,
            "title": "Principal",
            "type": "string"
          }
        },
        "required": [
          "auth_mode",
          "principal",
          "credential_ref"
        ],
        "title": "BrowseFoldersIn",
        "type": "object"
      },
      "BrowseFoldersOut": {
        "properties": {
          "folders": {
            "items": {
              "$ref": "#/components/schemas/DriveFolderOut"
            },
            "title": "Folders",
            "type": "array"
          },
          "parent_id": {
            "title": "Parent Id",
            "type": "string"
          }
        },
        "required": [
          "parent_id",
          "folders"
        ],
        "title": "BrowseFoldersOut",
        "type": "object"
      },
...
      "CredentialIn": {
        "description": "A source's credential. Write-only -- there is no read endpoint.",
        "properties": {
          "secret": {
            "description": "Service-account JSON key, or a JSON object with refresh_token, client_id and client_secret for OAuth.",
            "minLength": 1,
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret"
        ],
        "title": "CredentialIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunOut": {
        "properties": {
          "discovered": {
            "title": "Discovered",
            "type": "integer"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "ingested": {
            "title": "Ingested",
            "type": "integer"
          },
          "reconciled": {
            "title": "Reconciled",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "skips": {
            "items": {
              "$ref": "#/components/schemas/SkipOut"
            },
            "title": "Skips",
            "type": "array"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "unchanged": {
            "title": "Unchanged",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "source_id",
          "state",
          "started_at",
          "finished_at",
          "duration_seconds",
          "discovered",
          "ingested",
          "unchanged",
          "skipped",
          "failed",
          "reconciled",
          "error",
          "skips"
        ],
        "title": "RunOut",
        "type": "object"
      },
...
      "SourceIn": {
        "properties": {
          "auth_mode": {
            "enum": [
              "service_account",
              "oauth"
            ],
            "title": "Auth Mode",
            "type": "string"
          },
          "credential_ref": {
            "description": "Reference into the KMS-backed credential store. Never the credential itself.",
            "minLength": 1,
            "title": "Credential Ref",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "exclude_mime_types": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Exclude Mime Types",
            "type": "array"
          },
          "include_mime_types": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Include Mime Types",
            "type": "array"
          },
          "max_document_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Document Bytes"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "principal": {
            "description": "The effective identity ingestion acts as. Defines the corpus boundary and appears on every skip record.",
            "minLength": 1,
            "title": "Principal",
            "type": "string"
          },
          "provider": {
            "default": "google_drive",
            "title": "Provider",
            "type": "string"
          },
          "root_folder_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Root Folder Ids",
            "type": "array"
          }
        },
        "required": [
          "name",
          "auth_mode",
          "principal",
          "credential_ref"
        ],
        "title": "SourceIn",
        "type": "object"
      },
      "SourceOut": {
        "properties": {
          "auth_mode": {
            "enum": [
              "service_account",
              "oauth"
            ],
            "title": "Auth Mode",
            "type": "string"
          },
          "credential_ref": {
            "description": "Reference into the KMS-backed credential store. Never the credential itself.",
            "minLength": 1,
            "title": "Credential Ref",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "exclude_mime_types": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Exclude Mime Types",
            "type": "array"
          },
          "include_mime_types": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Include Mime Types",
            "type": "array"
          },
          "max_document_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Document Bytes"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "principal": {
            "description": "The effective identity ingestion acts as. Defines the corpus boundary and appears on every skip record.",
            "minLength": 1,
            "title": "Principal",
            "type": "string"
          },
          "provider": {
            "default": "google_drive",
            "title": "Provider",
            "type": "string"
          },
          "root_folder_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Root Folder Ids",
            "type": "array"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "auth_mode",
          "principal",
          "credential_ref",
          "source_id"
        ],
        "title": "SourceOut",
        "type": "object"
      },
      "StoredCredentialOut": {
        "description": "The result of storing a credential outside any source (ADR-0016).\n\nDecoupled from a source id on purpose: a credential can now be provisioned\n-- by pasting a key or by completing OAuth -- *before* the source it will\nbelong to has been created or saved, which is what makes browsing folders\nfrom the create dialog possible instead of requiring save-then-edit.",
        "properties": {
          "credential_ref": {
            "title": "Credential Ref",
            "type": "string"
          }
        },
        "required": [
          "credential_ref"
        ],
        "title": "StoredCredentialOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nRAG chat over Google Drive and other document sources.\n\n**Retrieval** is hybrid: dense pgvector search and Postgres full-text search are\nfused with Reciprocal Rank Fusion, then reranked (ADR-0004). Every turn streams\nits retrieval trace so a wrong answer can be diagnosed rather than merely\nnoticed (ADR-0009).\n\n**Permissions** are respected during ingestion. A folder that cannot be read is\nskipped *with an audit record* and never fails the run, because a silent skip is\nindistinguishable from an empty folder (ADR-0003).\n",
    "title": "RAGDrive API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/credentials": {
      "post": {
        "description": "Store a secret, encrypted at rest, under a freshly generated reference.\n\nThe reference is always server-generated, never typed by a user -- a\nhand-typed `credential_ref` is one typo from silently pointing a source at\nnothing, or at another source's key.",
        "operationId": "storeCredential",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredCredentialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Store Credential",
        "tags": [
          "credentials"
        ]
      }
    },
...
    "/sources": {
      "get": {
        "description": "Every registered source, enabled or not.",
        "operationId": "listSources",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceOut"
                  },
                  "title": "Response Listsources",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sources",
        "tags": [
          "sources"
        ]
      },
      "post": {
        "description": "Register a source.\n\n`credential_ref` points into the KMS-backed store; credential material never\nreaches this endpoint, so a request body in a log is not a disclosure.",
        "operationId": "createSource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Source",
        "tags": [
          "sources"
        ]
      }
    },
    "/sources/browse-folders": {
      "post": {
        "description": "List the subfolders of a Drive folder, for the config UI's folder picker.\n\nTakes a `credential_ref` directly rather than a source id, so this works\nfrom the create dialog too -- before any source row exists, right after\nOAuth connects or a service-account key is stored via `POST /credentials`.",
        "operationId": "browseFolders",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowseFoldersIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseFoldersOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Browse Folders",
        "tags": [
          "sources"
        ]
      }
    },
    "/sources/{source_id}": {
      "delete": {
        "description": "Remove a source and, by cascade, its documents and chunks.",
        "operationId": "deleteSource",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Source",
        "tags": [
          "sources"
        ]
      },
      "get": {
        "operationId": "getSource",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source",
        "tags": [
          "sources"
        ]
      },
      "put": {
        "description": "Replace a source's configuration in place.\n\nA full replace, not a partial patch: `PgSourceCatalogue.save` already\nupserts by id (it has since the source was first written), so the only\nthing missing was a route that called it a second time -- there was never a\npersistence-layer reason sources could only be created once.\n\n`source_id` is immutable; every other field, including `credential_ref`,\ncan change here. Changing `credential_ref` alone does not touch the secret\nit used to point at or the one it points at now -- pair this with\n`PUT /{source_id}/credential` (rotate what an existing ref holds) or a\nfresh `POST /credentials` (point at a different secret entirely).",
        "operationId": "updateSource",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Source",
        "tags": [
          "sources"
        ]
      }
    },
    "/sources/{source_id}/credential": {
      "put": {
        "description": "Store a source's credential, encrypted at rest.\n\nWrite-only by design. There is no GET: a credential that can be read back\nover HTTP is a credential one misconfigured permission away from disclosure,\nand nothing in the product needs to display it.",
        "operationId": "setSourceCredential",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Credential",
        "tags": [
          "sources"
        ]
      }
    },
    "/sources/{source_id}/ingest": {
      "post": {
        "description": "Start a run and return immediately.\n\nRuns take minutes to hours, so the request returns the run record and the\nclient polls `/runs/latest`. Holding the connection open for the duration\nwould fail behind any load balancer with an idle timeout.",
        "operationId": "startIngestion",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "incremental",
            "required": false,
            "schema": {
              "default": true,
              "title": "Incremental",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Ingestion",
        "tags": [
          "sources"
        ]
      }
    },
    "/sources/{source_id}/runs/latest": {
      "get": {
        "description": "The most recent run, including everything it could not read.\n\nThe skip list is the point (ADR-0003): a user needs to see what was excluded\nand why, because a silent skip is indistinguishable from an empty folder.",
        "operationId": "getLatestRun",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RunOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Getlatestrun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Run",
        "tags": [
          "sources"
        ]
      }
    },
...
  }
}